Fixed couple of issues in casem attachment indexer
